b-end/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
motor==3.6.0
pymongo==4.9.2
python-decouple==3.8
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.0.post1
dnspython==2.7.0

# Testing dependencies
pytest>=8.2.0
pytest-asyncio>=0.25.1
httpx==0.26.0
